refactor: separando a parte dos likes em outro componente
